track the web view's first completed loading cycle. Fixes #15
UIWebView’s `loading` is not a reliable source for detecting when a web
view has finished loading. It will continue to report `true` even when
the page appears to have finished loading. The best workaround I found
was to track the first time `loading` reports false in
webViewDidFinishLoad and consider that an estimation of a safe time to
unhide the web view.

This resolves issue #15, a bug that was causing the web view to remain
in a hidden state after making `presentModal()` and `animateForward()`
bridge calls.