stop loading and set web view delegate when disappearing from back button tap/pop
goBack is being called from the view controller that is disappearing.
because the web view’s delegate is not set until the transitioning view
controller’s viewWillAppear is called, the web view loads the previous
page from goBack() and the delegate methods are called on the
disappearing view controller. this erroneously causes the bridge API
object to point at the wrong view controller.

the fix is to stop loading the web view before calling back and set the
delegate to the top vc of the nav controller. after the delegate is
referencing the correct view controller goBack() can be called safely.